requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
PyPDF2
matplotlib
openai
requests
joblib
